pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-plugin-native-extract"
version = "0.1.0"
description = "Hermes plugin for native web content extraction — no API key required"
requires-python = ">= 3.11"
dependencies = [
    "requests>=2.28",
    "html-to-markdown",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pyyaml",
    "hermes-agent",
]
enhanced = [
    "certifi",
]

[project.entry-points."hermes_agent.plugins"]
native_extract_plugin = "__init__:register"

[tool.setuptools]
py-modules = ["__init__", "schemas", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
